// planning-idsdp-rms-frontend/src/app/main/modules/rpm/models/ResearcherProposalInstitutionInformation.ts
export class ResearcherProposalInstitutionInformation {

    id: number;
    uuid: string;
    researcherProposalId: number;
    instituteName: string;
    addressDetails: string;
    yearOfEstablishment: number;
    registrationNo: string;
    telephoneNo: string;
    emailAddress: string;
    tinNo: string;
    taxTokenImgOfCurrentYear: string;
    fileDownloadUrl: string;
    bucketName: string;
    fileName: string;
    isEditable: number;
}